requirements.txt
-e git://github.com/django/django.git#egg=Django
-e git://github.com/bendavis78/django-orderable.git#egg=django-orderable
-e git+git@git.farstarserver.com:libfarstar.git#egg=libfarstar
-e git://github.com/bendavis78/django-formhelper.git#egg=django-formhelper
South
django-tagging
textile
